fix(v3_4_3): translate mail packets to V3_4_3 wire format
Modern V3_4_3 client froze on mailbox open and rendered the inbox as
garbage (Enclosed amount: 1.69e+15, all attachments "Unknown"). Cause:
SMSG_MAIL_LIST_RESULT and the rest of the mail packet family used a
pre-V3_4_3 layout - Int32 MailID/AttachID, UInt8 SenderType, and a
sender-presence bit pair. CypherCore Source/Game/Networking/Packets/
MailPackets.cs:286-450 confirms V3_4_3 expects Int64 ids, UInt32
SenderType, and a switch-on-SenderType unconditional sender write.

Version-gated on ModernVersion.Build == V3_4_3_54261 to keep the
existing V1_14/V2_5 paths intact (vanilla send->list->take flow logged
clean in hermes-20260509_194751.log).

Affected packets: MailListEntry / MailAttachedItem / MailCommandResult
(SMSG writers) and MailDelete / MailMarkAsRead / MailReturnToSender /
MailTakeItem / MailTakeMoney / MailCreateTextItem (CMSG readers).
Server-side handler now casts (uint)mail.MailID etc. when forwarding
to the legacy server, since the storage fields widened to long.

+++ HermesProxy/World/Client/PacketHandlers/MailHandler.cs
@@ -102,7 +102,7 @@ public partial class WorldClient
                     GetSession().GameState.RequestedItemTextIds.Add(mail.ItemTextId);
                     WorldPacket query = new WorldPacket(Opcode.CMSG_ITEM_TEXT_QUERY);
                     query.WriteUInt32(mail.ItemTextId);
-                    query.WriteInt32(mail.MailID);
+                    query.WriteInt32((int)mail.MailID);
                     query.WriteUInt32(0); // unk
                     SendPacket(query);
                 }
